Rename updater ui to updater screen and some preparation for updater card
