+++ components/firstpair/impl/src/main/java/com/flipperdevices/firstpair/impl/composable/searching/ComposablePermissionRequest.kt
@@ -10,6 +10,7 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
@@ -52,7 +53,7 @@ fun ComposablePermissionRequest(modifier: Modifier, state: SearchingContent.Perm
         ComposableFlipperButton(
             modifier = Modifier.padding(all = 24.dp),
             text = stringResource(state.buttonText),
-            fontSize = 14.sp,
+            textStyle = TextStyle(fontSize = 14.sp),
             onClick = { state.onButtonClick() }
         )
     }
