+++ components/info/impl/src/main/java/com/flipperdevices/info/impl/compose/info/ComposableFullDeviceInfoScreen.kt
@@ -29,8 +29,10 @@ import com.flipperdevices.core.ktx.jre.titlecaseFirstCharIfItIsLowercase
 import com.flipperdevices.core.ui.ktx.LocalRouter
 import com.flipperdevices.core.ui.res.R as DesignSystem
 import com.flipperdevices.info.impl.R
-import com.flipperdevices.info.impl.compose.elements.InfoElementCard
 import com.flipperdevices.info.impl.viewmodel.DeviceInfoViewModel
+import com.flipperdevices.info.shared.ComposableDeviceInfoRowWithText
+import com.flipperdevices.info.shared.ComposableInfoDivider
+import com.flipperdevices.info.shared.InfoElementCard
 
 @Composable
 fun ComposableFullDeviceInfoScreen(
