+++ components/info/shared/src/main/res/values/strings.xml
@@ -9,4 +9,6 @@
     <string name="info_device_firmware_version_release">Release</string>
     <string name="info_device_firmware_version_release_full">Release</string>
     <string name="info_device_firmware_version_release_desc">Stable release (recommended)</string>
+
+    <string name="info_device_unknown">—</string>
 </resources>
