+++ components/updater/screen/src/main/res/drawable/ic_bluetooth.xml
