+++ components/updater/screen/src/main/res/values/colors.xml
