+++ components/updater/screen/src/main/res/values/strings.xml
