--- components/updater/ui/src/main/AndroidManifest.xml
@@ -1,1 +1,0 @@
-<manifest package="com.flipperdevices.updater.ui" />
