--- components/updater/ui/src/main/java/com/flipperdevices/updater/ui/di/UpdaterComponent.kt
@@ -1,14 +1,0 @@
-package com.flipperdevices.updater.ui.di
-
-import com.flipperdevices.core.di.AppGraph
-import com.flipperdevices.updater.ui.fragments.UpdaterFragment
-import com.flipperdevices.updater.ui.viewmodel.UpdateCardViewModel
-import com.flipperdevices.updater.ui.viewmodel.UpdaterViewModel
-import com.squareup.anvil.annotations.ContributesTo
-
-@ContributesTo(AppGraph::class)
-interface UpdaterComponent {
-    fun inject(viewModel: UpdateCardViewModel)
-    fun inject(viewModel: UpdaterViewModel)
-    fun inject(fragment: UpdaterFragment)
-}
